Add border width to box
