+++ src/components/RightSection/Box/Components/ChangeBorder.module.css
@@ -1,0 +1,43 @@
+.border {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 2rem;
+  border: 2px solid var(--color-text-black-light);
+  border-radius: 10px;
+  padding: 1rem 1.5rem;
+  margin-bottom: 1rem;
+  min-height: 11rem;
+  box-shadow: -6px 6px 2px 0px var(--color-background-dark-hard);
+}
+
+.border__label {
+  color: var(--color-text-black-light);
+}
+
+.border__input-container {
+  display: flex;
+  justify-content: space-between;
+  gap: 2rem;
+  width: 100%;
+}
+
+.border__span {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  color: var(--color-text-black-light);
+  font-size: 1.2rem;
+}
+
+.border__span-container {
+  position: relative;
+  border: 3px solid var(--color-text-black-light);
+  border-radius: 50%;
+  padding: 2rem;
+}
+
+.border__input {
+  width: 100%;
+}
